Fix minor typos and grammar
There's a bunch of ultra minor fixes in this checkin. Nothing that changes meaning.
Here are there I found run-on sentences and broke them up. Could've used a semicolon, but that's soooo academic :)
In a future checkin I'll make some suggestions for improvements in word usage, but first wait for other feedback that might alter the content.